refactor(preview): move findNextRunningBlockId pra helpers/
Eram ~60 linhas de funcao pura presas dentro do componente, sem usar nada de
React — fora da convencao do projeto e impossivel de testar isoladamente.
Agora vive em features/preview/helpers/, recebendo o typebot por parametro.
Pre-requisito dos testes pedidos no item B2 do review.

Endereca comentario de @martinezmn (A2):
https://github.com/cloudhumans/typebot.io/pull/272#pullrequestreview-4796605847
